[blog] binarylog: export Sink
So the sink can be updated (instead of only being noop)
